170 creating custom hook useMovies
!!!!!!!! Tova e s opciqta za zatvarqne na film, no taka ne e syvsem korektno. Tr da se doopravi neshto, no e slozhno na tozi etap. Zatova v sledvashtiq push shte e bez neq.

So there are basically two strategies to decide if you want to create a new custom hook. The first one is
that we want to reuse some part of our non-visual logic, so just as we learned in the previous lecture. And the
second factor may be that we simply want to extract a huge part of our component out into some custom hook.
And so that's actually what we will do in this lecture.
In this lecture it's shown to us how we can extract all the stateful logic that belongs together into a nice
and well-packaged custom hook.