193 starting a new quiz
In this way it's better.
If we were still using new state, then we would now probably create some new evend handler before the JSX in the App component and then we would pass that event handler into the start screen. But now, we don't need to do any of that anymore, because we are handling all the state transitions in the reducer. And then all we do is to pass the dispatcher functions around (to StartScreen) because then it's very easy. All we have to do here (in StartScreen component) is to call that dispatch function with the action that we want to dispatch ( () => onClick={() => dispatch... )